Rename mgmterrors package to aterrors
